Add random number, string hash and list helpers.
The random number use /dev/random if it can be open, or fallback to random()
otherwise.

The string hash use the MurmurHash as described on wikipedia. A common seed
is created from random number.

The list helpers provide the well know double linked list set of helpers to
add, remove and traverse a double linked list.

Signed-off-by: Jérôme Glisse <jglisse@redhat.com>